Fragment anidado en ActivityMain
Se ha creado un fragment anidado en el activity para poder ver otros fragments.
