+++ app/src/main/res/layout/activity_main.xml
@@ -8,20 +8,16 @@
     tools:context=".ui.MainActivity">
 
 
-        <androidx.recyclerview.widget.RecyclerView
-            android:id="@+id/recyclerview"
+        <fragment
+            android:id="@+id/navHostFragment"
             android:layout_width="match_parent"
             android:layout_height="0dp"
-            android:scrollbars="vertical"
-            app:layoutManager="androidx.recyclerview.widget.GridLayoutManager"
+            android:name="com.example.flickrgallery.ui.ExploreFragment"
             app:layout_constraintBottom_toTopOf="@+id/btnvMenu"
-            app:layout_constraintEnd_toEndOf="parent"
-            app:layout_constraintHorizontal_bias="0.0"
             app:layout_constraintStart_toStartOf="parent"
-            app:layout_constraintTop_toTopOf="parent"
-            app:layout_constraintVertical_bias="1.0"
-            app:spanCount="3"
-            tools:listitem="@layout/view_item_photo" />
+            app:layout_constraintTop_toTopOf="parent">
+
+        </fragment>
 
         <com.google.android.material.bottomnavigation.BottomNavigationView
             android:id="@+id/btnvMenu"
@@ -40,14 +36,14 @@
             android:id="@+id/floatingActionButton2"
             android:layout_width="wrap_content"
             android:layout_height="wrap_content"
-            android:layout_marginEnd="32dp"
-            android:layout_marginBottom="8dp"
+            android:layout_marginEnd="27dp"
+            android:layout_marginBottom="24dp"
             android:clickable="true"
+            android:contentDescription="@string/guardar_localizacion"
             app:backgroundTint="?android:attr/colorEdgeEffect"
-            app:layout_constraintBottom_toBottomOf="@+id/recyclerview"
-            app:layout_constraintEnd_toEndOf="@+id/recyclerview"
+            app:layout_constraintBottom_toTopOf="@+id/btnvMenu"
+            app:layout_constraintEnd_toEndOf="parent"
             app:maxImageSize="48dp"
-            app:srcCompat="@android:drawable/ic_menu_mylocation"
-            android:contentDescription="@string/guardar_localizacion" />
+            app:srcCompat="@android:drawable/ic_menu_mylocation" />
 
 </androidx.constraintlayout.widget.ConstraintLayout>
